fmh_bach29 change name test class to DocumentsControllerTest
