+++ backend/src/test/java/ru/iteco/fmh/service/DocumentsControllerTest.java
@@ -16,6 +16,15 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.context.junit4.SpringRunner;
+import ru.iteco.fmh.controller.DocumentsController;
+import ru.iteco.fmh.dao.repository.DocumentRepository;
+import ru.iteco.fmh.dto.document.DocumentCreationDtoRq;
+import ru.iteco.fmh.dto.document.DocumentCreationDtoRs;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static ru.iteco.fmh.TestUtils.*;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest()
@@ -37,5 +46,27 @@ public class DocumentsControllerTest {
                         .file(mockMultipartFile)
                         .contentType(MediaType.MULTIPART_FORM_DATA))
                 .andExpect(content().string(StringEndsWith.endsWith(".jpg")));
+    @Autowired
+    DocumentsController sut;
+    @Autowired
+    DocumentRepository documentRepository;
+    @Value("${static-host.host}")
+    private String staticHost;
+    @Test
+    public void createDocumentShouldPassSuccess() {
+        //given
+
+        DocumentCreationDtoRq givenDto = getDocumentCreationDtoRq();
+
+        DocumentCreationDtoRs resultDto = sut.createDocument(givenDto);
+
+        assertAll(
+                () -> assertEquals(givenDto.getName(), resultDto.getName()),
+                () -> assertEquals(staticHost+givenDto.getFilePath(), resultDto.getFilePath()),
+                () -> assertEquals(givenDto.getDescription(), resultDto.getDescription()),
+                () -> assertNotNull(resultDto.getId())
+        );
+        documentRepository.deleteById(resultDto.getId());
+
     }
 }
